Create adaptive_solver.py
Here is the first fully working version of the dimension-adaptive algorithm that we have all heard so much about. The code is as yet uncommented and fairly messy. Nevertheless it should suffice to give the reader a fair understanding of the main moving parts. If the global error stopping condition is set to '>0.2' it should have no trouble converging.